fixApi-Company-Major_v1

+++ OJT_RAG.Repositories/Interfaces/IJobDescriptionRepository.cs
@@ -6,7 +6,7 @@ namespace OJT_RAG.Repositories.Interfaces
     public interface IJobDescriptionRepository
     {
         Task<IEnumerable<JobDescription>> GetAll();
-        Task<JobDescription?> GetById(int id);
+        Task<JobDescription?> GetById(long id);
         Task<JobDescription> Add(JobDescription entity);
         Task<JobDescription> Update(JobDescription entity);
         Task<bool> Delete(int id);
